requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
pydantic-settings>=2.7
garminconnect>=0.2
alembic>=1.14
httpx>=0.28
fastapi-mcp>=0.2
pyyaml>=6.0
python-dotenv
psycopg2-binary>=2.9